+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+# 1.1.0
+## Content
+* Key Protect Support
+* Extended List Buckets Support
+### Defect Fixes
+* AWS Patches aligned to version 1.19.30 of the AWS SDK for Go
+
 # 1.0.1
 ## Content
 ### Defect Fixes